Complete backlog section 8: Cross-Platform & Build Improvements
- Add CMake as alternative build system (CMakeLists.txt + libft/CMakeLists.txt)
  with native Debug/Release build types and target-scoped strict warnings
- Fix all compiler warnings with stricter flags: -Wpedantic, -Wshadow,
  -Wconversion added to CFLAGS with -Werror (38 warnings fixed across 9 files)
- Add `make debug` target with ASan+UBSan (-fsanitize=address,undefined, -O0)
- Add `make release` target with -O3 -DNDEBUG (no debug symbols)
- Add multi-distro CI matrix: Ubuntu 22.04, Ubuntu 24.04, Debian 12, Fedora 39
- Document SDL2 migration path in docs/SDL2_MIGRATION.md (5 files affected,
  full API mapping provided)

+++ src/ft_bvh_2.c
@@ -43,7 +43,7 @@ t_bvh_node *ft_bvh_build(t_shape **shapes, int count)
 		node->box = ft_shape_aabb(shapes[0]);
 		return (node);
 	}
-	qsort(shapes, count, sizeof(t_shape *), ft_cmp_x);
+	qsort(shapes, (size_t)count, sizeof(t_shape *), ft_cmp_x);
 	mid = count / 2;
 	node->left = ft_bvh_build(shapes, mid);
 	node->right = ft_bvh_build(shapes + mid, count - mid);
@@ -124,7 +124,7 @@ void ft_build_scene_bvh(t_window *win)
 	}
 	if (count == 0)
 		return;
-	arr = malloc(sizeof(t_shape *) * count);
+	arr = malloc(sizeof(t_shape *) * (size_t)count);
 	if (!arr)
 		return;
 	cur = win->beg_sh;
